Modularise exporter state (#44175)
Fixes #42435.
