+++ client/state/selectors/get-media-export-url.js
@@ -3,4 +3,9 @@
  */
 import { get } from 'lodash';
 
+/**
+ * Internal dependencies
+ */
+import 'state/exporter/init';
+
 export default ( state ) => get( state, 'exporter.mediaExportUrl', null );
